refactor(domain): move api_versions.dart into domain/model/server/
